refactor some module package name
